super_user

+++ super_u/views.py
@@ -24,42 +24,58 @@ from django.contrib.auth.models import User
 def CreatePost(request):
   user = User 
   form = PostForm()
-  if request.method== 'POST':
-    post = Post(
-      title = request.POST['title'],
-      body  = request.POST['body'],
-      postImage = request.FILES['postImage'],
-      author = request.user
-    )
-    post.save()
-    return redirect('super_u:create_post')
+  if request.user.is_superuser:
+
+   if request.method== 'POST':
+     post = Post(
+       title = request.POST['title'],
+       body  = request.POST['body'],
+       postImage = request.FILES['postImage'],
+       author = request.user
+     )
+     post.save()
+     return redirect('super_u:create_post')
+  else:
+    return redirect('post:home')
   context = {'form':form}
 
   return render(request,'super_u/index.html',context)
 
 def listPosts(request):
-    posts = Post.objects.filter(author=request.user)
-    context = {'posts':posts}
+   if request.user.is_superuser:
+
+     posts = Post.objects.filter(author=request.user)
+     context = {'posts':posts}
+   else:
+    return redirect('post:home')
 
 
-    return render(request,'super_u/posts.html',context)
+   return render(request,'super_u/posts.html',context)
 
 
 def EditPost(request, id):
-  post = get_object_or_404(Post, id=id)
-  updateForm = PostForm(request.POST or None, instance=post)
-  if updateForm.is_valid():
-    updateForm.save()
-    return redirect('/super_u/posts.html')
-  context = {'updateForm':updateForm}
+  if request.user.is_superuser:
+
+   post = get_object_or_404(Post, id=id)
+   updateForm = PostForm(request.POST or None, instance=post)
+   if updateForm.is_valid():
+     updateForm.save()
+     return redirect('super_u:super_posts')
+   context = {'updateForm':updateForm}
+  else:
+    return redirect('post:home')
 
   return render(request,'super_u/update.html',context)
 
 def PostDelete(request, id):
-  post = get_object_or_404(Post, id=id)
-  if request.method == 'POST':
-    post.delete()
+  if request.user.is_superuser:
+
+   post = get_object_or_404(Post, id=id)
+   if request.method == 'POST':
+     post.delete()
     # return HttpResponse('super_u/posts.html')
-    return redirect('/super_u/post.html')
+     return redirect('super_u:super_posts')
+  else:
+    return redirect('post:home')
 
   return render(request,'super_u/posts.html')
